feat: move all code into the apps/palette-generator directory
